+++ test/rewards/initialize.spec.ts
@@ -1,5 +1,5 @@
 import { makeSuite, TestEnv } from '../helpers/make-suite';
-import { ZERO_ADDRESS } from '@mahalend/deploy-v3';
+import { ZERO_ADDRESS } from '@mahalend/deploy';
 
 const { expect } = require('chai');