+++ test/rewards/misc.spec.ts
@@ -4,7 +4,7 @@ import {
   MAX_UINT_AMOUNT,
   ZERO_ADDRESS,
   deployReservesSetupHelper,
-} from '@mahalend/deploy-v3';
+} from '@mahalend/deploy';
 import { expect } from 'chai';
 import { makeSuite } from '../helpers/make-suite';
 import { RANDOM_ADDRESSES } from '../helpers/constants';